rozbicie na main oraz na classes
