--- classes
@@ -1,1 +1,0 @@
-plik z klasami
